Fix BigDecimal to Integer type mismatches in price fields
Changed all price-related fields from BigDecimal to Integer to match the entity changes:

Entities:
- ServingStyle.additionalPrice: BigDecimal → Integer
- Order.totalPrice: BigDecimal → Integer
- OrderItem.price: BigDecimal → Integer

DTOs:
- CartResponse: all price fields → Integer
- OrderHistoryResponse.totalPrice → Integer
- OrderDetailResponse: totalPrice and price → Integer
- ActiveOrdersResponse.totalPrice → Integer

Services:
- CartService: Updated price calculations to use Integer arithmetic
- OrderService: Updated price calculations to use Integer arithmetic

This resolves all compilation errors related to type mismatches.

+++ src/main/java/com/ddoganzip/staff/dto/ActiveOrdersResponse.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -21,6 +20,6 @@ public class ActiveOrdersResponse {
     private LocalDateTime deliveryDate;
     private String deliveryAddress;
     private OrderStatus status;
-    private BigDecimal totalPrice;
+    private Integer totalPrice;
     private int itemCount;
 }
